Artworking browsing for staff and editing

+++ Museum-Backend/src/main/java/ca/mcgill/ecse321/museum/service/ArtworkService.java
@@ -198,11 +198,11 @@ public class ArtworkService {
     }
 
     // Edit artwork information
-    if (name != null)
+    if (name != null && name.trim().length() != 0)
       artwork.setName(name);
-    if (artist != null)
+    if (artist != null && artist.trim().length() != 0)
       artwork.setArtist(artist);
-    if (image != null)
+    if (image != null && image.trim().length() != 0)
       artwork.setImage(image);
     return artworkRepository.save(artwork);
   }
@@ -259,8 +259,8 @@ public class ArtworkService {
     }
 
     // Delete loan if artwork is on loan
-    Loan loan = loanRepository.findLoanByArtwork(artwork);
-    if (loan != null) {
+    List<Loan> loan = loanRepository.findLoanByArtwork(artwork);
+    if (loan != null && loan.size() > 0) {
       loanRepository.deleteLoanByArtwork(artwork);
     }
